block: use crash style time format for io pending time
