+++ drgn_tools/util.py
@@ -428,3 +428,16 @@ class BitNumberFlags(IntEnum):
         if not names or value:
             names.append(f"0x{value:x}")
         return "|".join(names)
+
+
+def timestamp_str(ns: int) -> str:
+    value = ns // 1000000
+    ms = value % 1000
+    value = value // 1000
+    secs = value % 60
+    value = value // 60
+    mins = value % 60
+    value = value // 60
+    hours = value % 24
+    days = value // 24
+    return "%d %02d:%02d:%02d.%03d" % (days, hours, mins, secs, ms)
